api-gateway/requirements.txt
Flask==3.1.0  
requests==2.31.0  
PyYAML==6.0  
